Update problem_data and sample_cases tables for upload
Uploading a problem now updates the sample_cases and problem_data tables
appropriately. Also included an api route that will return the JSON
representation of a single problem.

+++ auacm/app/modules/problem_manager/views.py
@@ -5,7 +5,7 @@ from app.database import Base, session
 from app.util import serve_response, serve_error, serve_info_pdf, login_manager
 from app.modules.user_manager.models import User
 from app.modules.submission_manager.models import Submission
-from app.modules.problem_manager.models import Problem
+from app.modules.problem_manager.models import Problem, Problem_Data, Sample_Case
 from os.path import join
 from sqlalchemy.orm import load_only
 from json import loads
@@ -20,11 +20,43 @@ def url_for_problem(problem):
 @login_required
 def get_problem_info(shortname):
     pid = session.query(Problem).\
-            options(load_only("pid", "shortname")).\
+            options(load_only('pid', 'shortname')).\
             filter(Problem.shortname==shortname).\
             first().pid
     return serve_info_pdf(str(pid))
 
+# Get a JSON representation of a problem
+@app.route('/api/problems/<shortname>', methods=['GET'])
+@login_required
+def get_problem(shortname):
+    problem = session.query(Problem).\
+            filter(Problem.shortname == shortname).\
+            first()
+    problem_data = session.query(Problem_Data).\
+                    filter(Problem_Data.pid  == problem.pid).\
+                    first()
+    case_results = session.query(Sample_Case).\
+                    filter(Sample_Case.pid == problem.pid).all()
+    cases = list()
+    for case in case_results:
+        cases.append({
+            'case_num': case.case_num,
+            'input': case.input,
+            'output': case.output
+        })
+    return serve_response({
+        'pid': problem.pid,
+        'name': problem.name,
+        'shortname': problem.shortname,
+        'appeared': problem.appeared,
+        'difficulty': problem.difficulty,
+        'added': problem.added,
+        'comp_release': problem.comp_release,
+        'description': problem_data.description,
+        'input_desc': problem_data.input_desc,
+        'output_desc': problem_data.output_desc,
+        'sample_cases': cases
+    })
 
 @app.route('/api/problems')
 @login_required
@@ -61,7 +93,7 @@ def create_problem():
         return serve_error('You must be an admin to create problems', 
             response_code=401)
 
-    # Ensure all parts of the form are complete (note js will send 'undefined')
+    # Ensure all required arguments are defined
     if not 'title' in request.form:
         return serve_error('You must give a problem title', response_code=400)
     if not 'description' in request.form:
@@ -76,30 +108,63 @@ def create_problem():
     if not 'cases' in request.form:
         return serve_error('You must provide at least one sample case',
             response_code=400)
-    # TODO(brandonlmorris): Check and handle zip files
-
-    # TODO(brandonlmorris): Insert new problem into database
-    title = request.form['title'][:32]
-    shortname = title.lower().replace(' ', '')
-
-    problem = Problem(
-        name=title,
-        shortname=shortname
-    )
-    pid = problem.commit_to_session()
 
     # Convert the JSON array string to python array of dictionaries
     cases = request.form['cases']
     cases = loads(str(cases))
+    for case in cases:
+        if not 'input' in case or not 'output' in case:
+            return serve_error('Sample case(s) were not formed correctly',
+                response_code=400)
+
+    # TODO(brandonlmorris): Check and handle zip files
+
+    # Create the problem and add it to the database
+    title = request.form['title'][:32]
+    shortname = title.lower().replace(' ', '')
+    problem = Problem(
+        name=title,
+        shortname=shortname
+    )
+    if 'difficulty' in request.form:
+        problem.diffiulty = request.form['diffiulty']
+    if 'appeared_in' in request.form:
+        problem.appeared = request.form['appeared_in']
+    pid = problem.commit_to_session()
+
+    # Create the problem data and add it to the database
+    problem_data = Problem_Data(
+        pid=pid,
+        description=request.form['description'],
+        input_desc=request.form['input_description'],
+        output_desc=request.form['output_description']
+    )
+    if 'time_limit' in request.form:
+        problem_data.time_limit = request.form['time_limit']
+    problem_data.commit_to_session()
+
+    # Add each sample case to the db
+    case_num = 1
+    for case in cases:
+        sample = Sample_Case(
+            pid=pid,
+            case_num=case_num,
+            input=case['input'],
+            output=case['output']
+        )
+        sample.commit_to_session()
+        case_num += 1
 
     return serve_response({
         'success': True,
-        'title': request.form['title'],
-        'description': request.form['description'],
-        'input_description': request.form['input_description'],
-        'output_description': request.form['output_description'],
+        'name': problem.name,
+        'shortname': problem.shortname,
+        'description': problem_data.description,
+        'input_description': problem_data.input_desc,
+        'output_description': problem_data.output_desc,
         'sample_cases': cases,
-        'pid': pid
+        'pid': problem.pid,
+        'difficulty': problem.difficulty
     })
 
 @app.route('/api/problems/delete', methods=['POST'])
@@ -114,13 +179,25 @@ def delete_problem():
         return serve_error('You must specify a problem to delete',
             response_code=400)
 
-    # TODO(brandonlmorris): Actually delete the thing
+    # Delete from problem_data table first to satisfy foreign key constraint
+    problem_data = session.query(Problem_Data).\
+        filter(Problem_Data.pid == request.form['pid'])
+    if not problem_data.first():
+        return serve_error('Could not find problem data with pid ' + 
+            request.form['pid'], response_code=401)
+    problem_data.delete()
+
+    # TODO(brandonlmorris):Delete any and all sample cases associated w/ problem
+
+    # Delete from problem table
     problem = session.query(Problem).\
         filter(Problem.pid == request.form['pid'])
     if not problem.first():
         return serve_error('Could not find problem with pid ' + 
             request.form['pid'], response_code=401)
     problem.delete()
+
+    # Commit changes
     session.flush()
     session.commit()
